[REEF-621] Set relaxLocality flag to ResourceRequestEvent objects
Currently, the relax locality flag is not being set when requesting resources.
Therefore, YarnResourceRequestHandler always sets it to true.
This change adds the flag when constructing the ResourceRequestEvent object.
In order to be backwards compatible, by default the flag will be set to true.
If the user defined rack names, and if none of those racks is the ANY modifier *, we will set it to false.

JIRA:
  [REEF-621](https://issues.apache.org/jira/browse/REEF-621)

Pull Request:
  Closes #